requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
fake-useragent==1.4.0
